+++ src/qrisp/jasp/tracing_logic/qaching.py
@@ -292,9 +292,9 @@ def qache_helper(func, jax_kwargs):
             if isinstance(args[i], bool):
                 args[i] = jnp.array(args[i], dtype = jnp.bool)
             elif isinstance(args[i], int):
-                args[i] = jnp.array(args[i], dtype = jnp.int32)
+                args[i] = jnp.array(args[i], dtype = jnp.int64)
             elif isinstance(args[i], float):
-                args[i] = jnp.array(args[i], dtype = jnp.float32)
+                args[i] = jnp.array(args[i], dtype = jnp.float64)
             elif isinstance(args[i], complex):
                 args[i] = jnp.array(args[i], dtype = jnp.complex)
